requirements.txt
openai>=1.40.0
pytest>=7.0.0
anthropic>=0.7.0
pyautogen>=0.2.0
google-generativeai
autogen-agentchat[gemini]~=0.2
